requirements.txt
flask==1.1.1
flask-redis==0.4.0
requests==2.22.0
python-dotenv==0.10.3